+++ calculations/calculator.go
@@ -1,4 +1,4 @@
-package business
+package calculations
 
 import "errors"
